requirements.txt
fastapi==0.115.10
uvicorn==0.34.0
telethon==1.39.0
python-dotenv==1.0.1
websockets==15.0
google-generativeai==0.8.2
ccxt==4.4.15
